+++ src/main/java/org/mastodon/revised/model/mamut/trackmate/TrackMateImporter.java
@@ -1,28 +1,28 @@
-package org.mastodon.revised.model.mamut.tm2;
+package org.mastodon.revised.model.mamut.trackmate;
 
-import static org.mastodon.revised.model.mamut.tm2.TrackMateXMLKeys.EDGE_FEATURE_DECLARATION_TAG;
-import static org.mastodon.revised.model.mamut.tm2.TrackMateXMLKeys.EDGE_SOURCE_ATTRIBUTE_NAME;
-import static org.mastodon.revised.model.mamut.tm2.TrackMateXMLKeys.EDGE_TAG;
-import static org.mastodon.revised.model.mamut.tm2.TrackMateXMLKeys.EDGE_TARGET_ATTRIBUTE_NAME;
-import static org.mastodon.revised.model.mamut.tm2.TrackMateXMLKeys.FEATURE_ATTRIBUTE_NAME;
-import static org.mastodon.revised.model.mamut.tm2.TrackMateXMLKeys.FEATURE_DECLARATION_TAG;
-import static org.mastodon.revised.model.mamut.tm2.TrackMateXMLKeys.FEATURE_ISINT_ATTRIBUTE_NAME;
-import static org.mastodon.revised.model.mamut.tm2.TrackMateXMLKeys.FEATURE_TAG;
-import static org.mastodon.revised.model.mamut.tm2.TrackMateXMLKeys.FRAME_ATTRIBUTE_NAME;
-import static org.mastodon.revised.model.mamut.tm2.TrackMateXMLKeys.ID_ATTRIBUTE_NAME;
-import static org.mastodon.revised.model.mamut.tm2.TrackMateXMLKeys.LABEL_ATTRIBUTE_NAME;
-import static org.mastodon.revised.model.mamut.tm2.TrackMateXMLKeys.MODEL_TAG;
-import static org.mastodon.revised.model.mamut.tm2.TrackMateXMLKeys.RADIUS_ATTRIBUTE_NAME;
-import static org.mastodon.revised.model.mamut.tm2.TrackMateXMLKeys.SPOT_COLLECTION_TAG;
-import static org.mastodon.revised.model.mamut.tm2.TrackMateXMLKeys.SPOT_ELEMENT_TAG;
-import static org.mastodon.revised.model.mamut.tm2.TrackMateXMLKeys.SPOT_FEATURE_DECLARATION_TAG;
-import static org.mastodon.revised.model.mamut.tm2.TrackMateXMLKeys.SPOT_FRAME_COLLECTION_TAG;
-import static org.mastodon.revised.model.mamut.tm2.TrackMateXMLKeys.TRACK_COLLECTION_TAG;
-import static org.mastodon.revised.model.mamut.tm2.TrackMateXMLKeys.TRACK_TAG;
-import static org.mastodon.revised.model.mamut.tm2.TrackMateXMLKeys.VISIBILITY_ATTRIBUTE_NAME;
-import static org.mastodon.revised.model.mamut.tm2.TrackMateXMLKeys.X_ATTRIBUTE_NAME;
-import static org.mastodon.revised.model.mamut.tm2.TrackMateXMLKeys.Y_ATTRIBUTE_NAME;
-import static org.mastodon.revised.model.mamut.tm2.TrackMateXMLKeys.Z_ATTRIBUTE_NAME;
+import static org.mastodon.revised.model.mamut.trackmate.TrackMateXMLKeys.EDGE_FEATURE_DECLARATION_TAG;
+import static org.mastodon.revised.model.mamut.trackmate.TrackMateXMLKeys.EDGE_SOURCE_ATTRIBUTE_NAME;
+import static org.mastodon.revised.model.mamut.trackmate.TrackMateXMLKeys.EDGE_TAG;
+import static org.mastodon.revised.model.mamut.trackmate.TrackMateXMLKeys.EDGE_TARGET_ATTRIBUTE_NAME;
+import static org.mastodon.revised.model.mamut.trackmate.TrackMateXMLKeys.FEATURE_ATTRIBUTE_NAME;
+import static org.mastodon.revised.model.mamut.trackmate.TrackMateXMLKeys.FEATURE_DECLARATION_TAG;
+import static org.mastodon.revised.model.mamut.trackmate.TrackMateXMLKeys.FEATURE_ISINT_ATTRIBUTE_NAME;
+import static org.mastodon.revised.model.mamut.trackmate.TrackMateXMLKeys.FEATURE_TAG;
+import static org.mastodon.revised.model.mamut.trackmate.TrackMateXMLKeys.FRAME_ATTRIBUTE_NAME;
+import static org.mastodon.revised.model.mamut.trackmate.TrackMateXMLKeys.ID_ATTRIBUTE_NAME;
+import static org.mastodon.revised.model.mamut.trackmate.TrackMateXMLKeys.LABEL_ATTRIBUTE_NAME;
+import static org.mastodon.revised.model.mamut.trackmate.TrackMateXMLKeys.MODEL_TAG;
+import static org.mastodon.revised.model.mamut.trackmate.TrackMateXMLKeys.RADIUS_ATTRIBUTE_NAME;
+import static org.mastodon.revised.model.mamut.trackmate.TrackMateXMLKeys.SPOT_COLLECTION_TAG;
+import static org.mastodon.revised.model.mamut.trackmate.TrackMateXMLKeys.SPOT_ELEMENT_TAG;
+import static org.mastodon.revised.model.mamut.trackmate.TrackMateXMLKeys.SPOT_FEATURE_DECLARATION_TAG;
+import static org.mastodon.revised.model.mamut.trackmate.TrackMateXMLKeys.SPOT_FRAME_COLLECTION_TAG;
+import static org.mastodon.revised.model.mamut.trackmate.TrackMateXMLKeys.TRACK_COLLECTION_TAG;
+import static org.mastodon.revised.model.mamut.trackmate.TrackMateXMLKeys.TRACK_TAG;
+import static org.mastodon.revised.model.mamut.trackmate.TrackMateXMLKeys.VISIBILITY_ATTRIBUTE_NAME;
+import static org.mastodon.revised.model.mamut.trackmate.TrackMateXMLKeys.X_ATTRIBUTE_NAME;
+import static org.mastodon.revised.model.mamut.trackmate.TrackMateXMLKeys.Y_ATTRIBUTE_NAME;
+import static org.mastodon.revised.model.mamut.trackmate.TrackMateXMLKeys.Z_ATTRIBUTE_NAME;
 
 import java.io.File;
 import java.io.IOException;
